baio_workbench/requirements.txt
numpy-groupies==0.10.2
pandas==2.1.1
scipy==1.11.3
biopython==1.81
swig==4.1.1
swalign==0.3.7
scanpy==1.9.6
vireo==0.15.1
kaleido==0.2.1
plotly==5.18.0
langchain
langchain-community==0.0.2
langchain-core
langchain-experimental
langchainhub
openai==0.28.1
pydantic_core==2.10.1
requests-oauthlib==1.3.1
streamlit==1.28.0
pypdfium2==4.23.1
chromadb==0.3.29
tokenizers==0.14.1
tiktoken==0.5.1
rdflib==7.0.0
lxml==4.9.3
unstructured==0.10.28
layoutparser==0.3.4
mygene==3.2.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
xmltodict